# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Yang's Fried Rice - System API",
    "version": "1.0.0",
    "description": "Interactive Swagger/OpenAPI documentation for Menu Management, Administration, Order Oversight, and Reporting. Use 'Try it out' to execute live requests."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get all remote orders",
        "description": "Lists orders with joined customer name, item details, delivery info, and transaction status. Filterable by status and date range. Requires admin, manager, or staff.",
        "operationId": "getOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "received",
                "preparing",
                "out_for_delivery",
                "completed",
                "cancelled"
              ]
            },
            "description": "Filter by order status"
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter orders created on or after this date (YYYY-MM-DD)"
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter orders created on or before this date (YYYY-MM-DD)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Pagination limit"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Pagination offset"
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 5
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/stats": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order status statistics",
        "description": "Returns order counts grouped by status (received, preparing, out_for_delivery, completed, cancelled, total). Requires admin, manager, or staff.",
        "operationId": "getOrderStats",
        "responses": {
          "200": {
            "description": "Order statistics counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OrderStats"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get order details",
        "description": "Retrieves complete order detail including customer info, product line items, payment transaction, and delivery record. Requires admin, manager, or staff.",
        "operationId": "getOrderDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Order ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}/status": {
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "Update order status",
        "description": "Updates order status through stages: received -> preparing -> out_for_delivery -> completed (or cancelled from any non-completed stage). Requires admin, manager, or staff.",
        "operationId": "updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Order ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStatusUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order status updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status transition or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Internal server error"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "Order": {
        "type": "object",
        "required": [
          "order_id",
          "customer_id",
          "order_status"
        ],
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid",
            "example": "o0000000-0000-0000-0000-000000000001"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "example": "c0000000-0000-0000-0000-000000000001"
          },
          "order_status": {
            "type": "string",
            "enum": [
              "received",
              "preparing",
              "out_for_delivery",
              "completed",
              "cancelled"
            ],
            "example": "received"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-10T09:00:00Z"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "cancelled_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "special_instructions": {
            "type": "string",
            "nullable": true,
            "example": "Less oil please"
          }
        }
      },
      "OrderStatusUpdateInput": {
        "type": "object",
        "required": [
          "new_status"
        ],
        "properties": {
          "new_status": {
            "type": "string",
            "enum": [
              "received",
              "preparing",
              "out_for_delivery",
              "completed",
              "cancelled"
            ],
            "example": "preparing"
          }
        }
      },
      "OrderStats": {
        "type": "object",
        "required": [
          "received",
          "preparing",
          "out_for_delivery",
          "completed",
          "cancelled",
          "total"
        ],
        "properties": {
          "received": {
            "type": "integer",
            "example": 4
          },
          "preparing": {
            "type": "integer",
            "example": 2
          },
          "out_for_delivery": {
            "type": "integer",
            "example": 3
          },
          "completed": {
            "type": "integer",
            "example": 45
          },
          "cancelled": {
            "type": "integer",
            "example": 1
          },
          "total": {
            "type": "integer",
            "example": 55
          }
        }
      },
...
    }
  }
}